# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.Ticket:
    properties:
      assigned_to_id:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      created_by_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      metadata:
        description: JSONB field for custom metadata
        type: string
      notes:
        type: string
      priority:
        $ref: '#/definitions/models.TicketPriority'
      related_asset_id:
        type: integer
      status:
        $ref: '#/definitions/models.TicketStatus'
      tenant_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TicketPriority:
    enum:
    - low
    - medium
    - high
    - critical
    type: string
    x-enum-varnames:
    - TicketPriorityLow
    - TicketPriorityMedium
    - TicketPriorityHigh
    - TicketPriorityCritical
  models.TicketStatus:
    enum:
    - open
    - in_progress
    - resolved
    - closed
    type: string
    x-enum-varnames:
    - TicketStatusOpen
    - TicketStatusInProgress
    - TicketStatusResolved
    - TicketStatusClosed
...
host: localhost:3000
info:
  contact:
    email: support@kontena.io
    name: API Support
  description: Multi-tenant CRM API built with Golang and Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kontena CRM API
  version: "1.0"
paths:
...
  /tenants:
    get:
      consumes:
      - application/json
      description: Get all tenants in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
      summary: Get all tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant
      parameters:
      - description: Tenant object
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tenant
      tags:
      - tenants
    get:
      consumes:
      - application/json
      description: Get a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a tenant
      tags:
      - tenants
    put:
      consumes:
      - application/json
      description: Update a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tenant object
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a tenant
      tags:
      - tenants
  /tickets:
    get:
      consumes:
      - application/json
      description: Get all tickets for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ticket'
            type: array
      summary: Get all tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Create a new ticket for the current tenant
      parameters:
      - description: Ticket information
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/models.Ticket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a ticket
      tags:
      - tickets
  /tickets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a ticket by ID for the current tenant
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a ticket
      tags:
      - tickets
    get:
      consumes:
      - application/json
      description: Get a ticket by ID for the current tenant
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a ticket
      tags:
      - tickets
    patch:
      consumes:
      - application/json
      description: Update an existing ticket for the current tenant
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket information
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/models.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a ticket
      tags:
      - tickets
...
swagger: "2.0"